Revert "Moved Application tab functionality to separate class"
